+++ src/Jadob/Core/Kernel.php
@@ -56,7 +56,7 @@ class Kernel
      * @see https://semver.org/
      * @var string
      */
-    public const VERSION = '0.2.3';
+    public const VERSION = '0.2.4';
 
     /**
      * If true, application log will be saved while destructing objects