+++ CHANGELOG.md
@@ -4,12 +4,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## Unreleased
+## [0.2.4] - 2021-12-13
 ### Added
 - [Container] added support for parameter
 - [Container] injecting parameters to services via `InjecParam` attribute
 - [Dashboard] Registered more Twig extensions
 - Adding parameters in configuration
+
+### Changed
+- bumped php version
 
 ## [0.2.3] - 2021-09-17
 ### Added